demo_trailing_stops.py
"""
Trailing Stop-Loss Demo & Testing
=================================

This script demonstrates and tests the comprehensive trailing stop-loss
functionality integrated into your trading system.
"""

import sys
from pathlib import Path
import pandas as pd
import numpy as np
from datetime import datetime, timedelta
import matplotlib.pyplot as plt

# Add src to path
sys.path.insert(0, str(Path(__file__).parent / "src"))

def demo_trailing_stops():
    """Demonstrate trailing stop functionality."""
    
    print("🎯 TRAILING STOP-LOSS SYSTEM DEMO")
    print("=" * 60)
    print("Testing comprehensive trailing stop functionality...")
    
    # Simulate a profitable trade with trailing stops
    print(f"\n📊 SCENARIO: Profitable Trade with Trailing Stops")
    print("-" * 50)
    
    # Trade parameters
    entry_price = 1000.0
    initial_stop_loss = 920.0  # 8% stop loss
    take_profit = 1160.0      # 16% take profit
    trailing_stop_pct = 0.04  # 4% trailing stop
    
    print(f"💰 Entry Price: ₹{entry_price:.2f}")
    print(f"🛡️  Initial Stop Loss: ₹{initial_stop_loss:.2f} (-{((entry_price - initial_stop_loss) / entry_price * 100):.1f}%)")
    print(f"🎯 Take Profit: ₹{take_profit:.2f} (+{((take_profit - entry_price) / entry_price * 100):.1f}%)")
    print(f"📈 Trailing Stop: {trailing_stop_pct * 100:.1f}% below highest price")
    
    # Simulate price movements
    days = 10
    np.random.seed(42)  # For reproducible results
    
    # Generate realistic price movement (trending up with volatility)
    price_changes = np.random.normal(0.01, 0.03, days)  # 1% daily trend, 3% volatility
    prices = [entry_price]
    
    for change in price_changes:
        new_price = prices[-1] * (1 + change)
        prices.append(max(new_price, prices[-1] * 0.95))  # Prevent huge drops
    
    # Add a final drop to test trailing stop
    prices.extend([prices[-1] * 0.98, prices[-1] * 0.96, prices[-1] * 0.94])
    
    # Calculate trailing stops
    highest_price = entry_price
    trailing_stop = None
    trailing_stops = []
    
    print(f"\n📈 DAY-BY-DAY PRICE MOVEMENT & TRAILING STOPS")
    print("-" * 60)
    print(f"{'Day':<4} {'Price':<10} {'Highest':<10} {'Trail Stop':<12} {'Status':<15}")
    print("-" * 60)
    
    for day, price in enumerate(prices):
        # Update highest price
        if price > highest_price:
            highest_price = price
        
        # Update trailing stop (only when in profit)
        if highest_price > entry_price:
            new_trailing_stop = highest_price * (1 - trailing_stop_pct)
            if trailing_stop is None:
                trailing_stop = max(initial_stop_loss, new_trailing_stop)
            else:
                trailing_stop = max(trailing_stop, new_trailing_stop)
        
        trailing_stops.append(trailing_stop)
        
        # Check exit conditions
        status = "ACTIVE"
        if trailing_stop and price <= trailing_stop:
            status = "TRAIL STOP HIT"
        elif price <= initial_stop_loss:
            status = "STOP LOSS HIT"
        elif price >= take_profit:
            status = "TAKE PROFIT HIT"
        
        print(f"{day:<4} ₹{price:<9.2f} ₹{highest_price:<9.2f} ₹{trailing_stop if trailing_stop else 'None':<11} {status:<15}")
        
        # Exit if stop hit
        if "HIT" in status:
            exit_price = price
            exit_reason = status
            break
    
    # Calculate final result
    if 'exit_price' in locals():
        pnl = exit_price - entry_price
        pnl_pct = (pnl / entry_price) * 100
        
        print(f"\n🎯 TRADE RESULT")
        print("-" * 30)
        print(f"💰 Entry Price: ₹{entry_price:.2f}")
        print(f"🚪 Exit Price: ₹{exit_price:.2f}")
        print(f"📊 Exit Reason: {exit_reason}")
        print(f"💹 P&L: ₹{pnl:.2f} ({pnl_pct:+.2f}%)")
        
        if "TRAIL" in exit_reason:
            print(f"✅ Trailing stop protected profits!")
            print(f"🛡️  Protected: ₹{exit_price - entry_price:.2f} profit")
        elif pnl > 0:
            print(f"✅ Profitable trade!")
        else:
            print(f"❌ Loss contained by stop loss")


def test_backtesting_with_trailing_stops():
    """Test backtesting with trailing stops enabled."""
    
    print(f"\n🧪 BACKTESTING WITH TRAILING STOPS")
    print("=" * 60)
    
    try:
        from trading_system.config import TradingConfig
        from trading_system.data_manager import DataManager
        from trading_system.technical_analysis import TechnicalAnalyzer
        from trading_system.risk_manager import RiskManager
        from trading_system.backtester import Backtester
        
        # Initialize system
        config = TradingConfig()
        data_manager = DataManager(config)
        technical_analyzer = TechnicalAnalyzer(config)
        risk_manager = RiskManager(config)
        backtester = Backtester(config, data_manager, technical_analyzer, risk_manager)
        
        # Test with a few liquid stocks
        test_stocks = ['RELIANCE.NS', 'TCS.NS', 'HDFCBANK.NS', 'INFY.NS', 'ICICIBANK.NS']
        
        print(f"📊 Testing trailing stops on {len(test_stocks)} stocks")
        print("🔄 Running 2-month backtest with trailing stops...")
        
        # Run short backtest
        end_date = datetime.now().strftime('%Y-%m-%d')
        start_date = (datetime.now() - timedelta(days=60)).strftime('%Y-%m-%d')
        
        metrics = backtester.run_backtest(
            symbols=test_stocks,
            start_date=start_date,
            end_date=end_date,
            initial_capital=100000
        )
        
        # Get trade history to analyze trailing stops
        trade_history = backtester.get_trade_history()
        
        print(f"\n📊 TRAILING STOP ANALYSIS")
        print("-" * 50)
        print(f"Total Trades: {len(trade_history)}")
        
        if not trade_history.empty:
            # Analyze trailing stop usage
            trailing_stop_exits = trade_history[trade_history['Exit Reason'] == 'TRAILING_STOP']
            print(f"Trailing Stop Exits: {len(trailing_stop_exits)}")
            
            if len(trailing_stop_exits) > 0:
                avg_pnl = trailing_stop_exits['P&L (%)'].mean()
                print(f"Avg P&L from Trailing Stops: {avg_pnl:.2f}%")
                
                print(f"\n🎯 TRAILING STOP TRADES:")
                for _, trade in trailing_stop_exits.iterrows():
                    print(f"• {trade['Symbol']}: {trade['P&L (%)']:+.2f}% (₹{trade['Entry Price']:.2f} → ₹{trade['Exit Price']:.2f})")
            
            # Show detailed trade with trailing stop info
            if len(trade_history) > 0:
                print(f"\n📋 SAMPLE TRADE WITH TRAILING STOP DETAILS:")
                sample_trade = trade_history.iloc[0]
                print(f"Symbol: {sample_trade['Symbol']}")
                print(f"Entry: ₹{sample_trade['Entry Price']:.2f}")
                print(f"Stop Loss: ₹{sample_trade['Stop Loss']:.2f}")
                print(f"Take Profit: ₹{sample_trade['Take Profit']:.2f}")
                if pd.notna(sample_trade['Trailing Stop']):
                    print(f"Final Trailing Stop: ₹{sample_trade['Trailing Stop']:.2f}")
                if pd.notna(sample_trade['Highest Price']):
                    print(f"Highest Price Reached: ₹{sample_trade['Highest Price']:.2f}")
                print(f"Exit: ₹{sample_trade['Exit Price']:.2f} ({sample_trade['Exit Reason']})")
        else:
            print("No trades executed in test period.")
            print("💡 Try running: python run_backtest.py for comprehensive testing")
        
    except Exception as e:
        print(f"❌ Error testing trailing stops: {e}")
        print("💡 Make sure to run this from the project directory")


def show_trailing_stop_configuration():
    """Show current trailing stop configuration."""
    
    print(f"\n⚙️ TRAILING STOP CONFIGURATION")
    print("=" * 60)
    
    try:
        from trading_system.config import TradingConfig
        config = TradingConfig()
        
        print(f"📊 Current Settings:")
        print(f"• Trailing Stop Percentage: {config.risk.trailing_stop_pct * 100:.1f}%")
        print(f"• Risk Per Trade: {config.risk.risk_per_trade * 100:.1f}%")
        print(f"• Stop Loss Percentage: {config.risk.stop_loss_pct * 100:.1f}%")
        print(f"• Take Profit Percentage: {config.risk.take_profit_pct * 100:.1f}%")
        
        print(f"\n🎯 How Trailing Stops Work:")
        print("1. Initial stop loss protects against immediate losses")
        print("2. When price moves in your favor, trailing stop activates")
        print("3. Trailing stop follows price up, maintaining 4% distance")
        print("4. If price drops 4% from highest point, position exits")
        print("5. This locks in profits while allowing for continued upside")
        
        print(f"\n💡 Example with ₹1000 stock:")
        entry = 1000
        trail_pct = config.risk.trailing_stop_pct
        
        scenarios = [
            (1050, "Stock rises to ₹1050"),
            (1100, "Stock continues to ₹1100"),
            (1056, "Stock drops to ₹1056 (4% from ₹1100)")
        ]
        
        highest = entry
        for price, description in scenarios:
            if price > highest:
                highest = price
            trailing_stop = highest * (1 - trail_pct) if highest > entry else None
            
            print(f"• {description}")
            print(f"  Highest so far: ₹{highest:.2f}")
            print(f"  Trailing stop: ₹{trailing_stop:.2f}" if trailing_stop else "  Trailing stop: Not active yet")
            
            if trailing_stop and price <= trailing_stop:
                profit = price - entry
                print(f"  🚨 TRAILING STOP TRIGGERED! Profit locked: ₹{profit:.2f}")
                break
        
    except Exception as e:
        print(f"Error loading configuration: {e}")


def main():
    """Main demo function."""
    print("🎯 COMPREHENSIVE TRAILING STOP-LOSS SYSTEM")
    print("=" * 70)
    print("Your trading system includes advanced trailing stop functionality!")
    print("=" * 70)
    
    # Show configuration
    show_trailing_stop_configuration()
    
    # Demo trailing stops
    demo_trailing_stops()
    
    # Test with backtesting
    test_backtesting_with_trailing_stops()
    
    print(f"\n✅ TRAILING STOP INTEGRATION STATUS")
    print("=" * 60)
    print("✅ Risk Manager: Trailing stops implemented")
    print("✅ Backtester: Enhanced with trailing stop tracking")
    print("✅ Configuration: 4% trailing stop percentage")
    print("✅ Trade History: Includes trailing stop details")
    print("✅ Exit Reasons: Tracks trailing stop exits")
    
    print(f"\n🚀 READY FOR LIVE TRADING WITH TRAILING STOPS!")
    print("Run your analysis commands to see trailing stops in action:")
    print("• python analyze_today.py")
    print("• python run_backtest.py")
    print("• python main.py backtest")


if __name__ == "__main__":
    main()
