requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
ta>=0.10.2
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
schedule>=1.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
streamlit>=1.29.0
sqlalchemy>=2.0.0
python-telegram-bot>=20.7
discord.py>=2.3.0
